+++ src/style/style.test.ts
@@ -179,6 +179,20 @@ describe('Style.loadURL', () => {
         expect(error).toBeTruthy();
         expect(error.status).toBe(errorStatus);
     });
+
+    test('does not throw if request is pending when removed', async () => {
+        const style = new Style(getStubMap());
+        style.loadJSON(createStyleJSON());
+        await style.once('style.load');
+
+        const errorHandler = vi.fn();
+        style.on('error', errorHandler);
+
+        style.loadURL('style.json');
+        style._remove();
+
+        expect(errorHandler).not.toHaveBeenCalled();
+    });
 });
 
 describe('Style.loadJSON', () => {
@@ -534,6 +548,26 @@ describe('Style.loadJSON', () => {
         const paint = layer.paint as PossiblyEvaluated<CirclePaintProps, CirclePaintPropsPossiblyEvaluated>;
         expect(paint.get('circle-color').evaluate({} as Feature, {})).toEqual(new Color(1, 0, 0, 1));
         expect(paint.get('circle-radius').evaluate({} as Feature, {})).toEqual(12);
+    });
+
+    test('does not throw if request is pending when removed', async () => {
+        const style = new Style(getStubMap());
+        style.loadJSON(createStyleJSON());
+        await style.once('style.load');
+
+        const errorHandler = vi.fn();
+        style.on('error', errorHandler);
+
+        style.loadJSON(
+            createStyleJSON({
+                sources: {
+                    'source-id': createGeoJSONSource()
+                },
+            })
+        );
+        style._remove();
+
+        expect(errorHandler).not.toHaveBeenCalled();
     });
 });
 
@@ -1306,6 +1340,22 @@ describe('Style.addSprite', () => {
         expect(inputJson.sprite).toBe(inputSprite);
         expect(JSON.stringify(inputJson)).toEqual(inputJsonString);
     });
+
+    test('does not throw if request is pending when removed', async () => {
+        const style = new Style(getStubMap());
+        style.loadJSON(createStyleJSON());
+        await style.once('style.load');
+
+        const errorHandler = vi.fn();
+        style.on('error', errorHandler);
+
+        style.addSprite('test', 'https://example.com/sprite');
+        style._remove();
+
+        await waitForEvent(style, 'data', (event) => event.dataType === 'style');
+
+        expect(errorHandler).not.toHaveBeenCalled();
+    });
 });
 
 describe('Style.removeSprite', () => {
@@ -1391,6 +1441,20 @@ describe('Style.setSprite', () => {
 
         expect(inputJson.sprite).toBe(inputSprite);
         expect(inputJsonString).toEqual(JSON.stringify(inputJson));
+    });
+
+    test('throws when error loading sprite', async () => {
+        server.respondWith('https://example.com/sprite', [404, {}, '']);
+        const style = new Style(getStubMap());
+        style.loadJSON(createStyleJSON());
+        await style.once('style.load');
+
+        const errorPromise = style.once('error');
+        style.setSprite('https://example.com/sprite');
+        server.respond();
+
+        const {error} = await errorPromise;
+        expect(error.message).toBe('AJAXError: Not Found (404): https://example.com/sprite.json');
     });
 });